requirements.txt
finnhub-python==2.4.19
rich==13.7.1
websockets
